==> Banco_de_Dados/06. Google BigQuery/3.Conexão com Auyth2 Variavel.ipynb <==
from google.cloud import bigquery
import pandas as pd
from google.oauth2 import service_account

# Especifique as informaçoes de credenciais
credencial = {
    
  "type": "service_account",
  "project_id": "pythonengenharia",
  "private_key_id": "e4145aaaeb10a3",
  "private_key": "-----pythonengenhariaprojeto@pythonengenharia.iam.gserviceaccount.com",
  "client_id": "10068232482",
  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
  "token_uri": "https://oauth2.googleapis.com/token",
  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/pythonengenhariaprojeto%40pythonengenharia.iam.gserviceaccount.com",
 
}


# Crie o cliente BigQuery
client = bigquery.Client(credentials=service_account.Credentials.from_service_account_info(credencial))

# Construa sua consulta
query = """ 

  SELECT 
  id
  ,first_name
  ,last_name
  ,email  
  ,state
 FROM `pythoncurso-398511.AulaPython.Clientes` 
 where state in ('Piauí','Alagoas',
 'Amazonas','Maranhão','Rondônia')
"""
# Execute a consulta
resultado = client.query(query)

# Converta os resultados em um DataFrame
df = resultado.to_dataframe()
df.head(30)
