==> Banco_de_Dados/06. Google BigQuery/4.Carregando Dados para o Google BigQuery.ipynb <==
#pip install google-cloud-bigquery
#pip install --upgrade google-cloud-bigquery
#pip install pandas

#pip install pandas pandas-gbq

import pandas as pd
from google.cloud import bigquery
from google.oauth2 import service_account
from pandas_gbq import to_gbq

# Carregue suas credenciais
credencial = service_account.Credentials.from_service_account_file(
    r'C:\Users\Note reserva\Downloads\pythonengenharia-e4145aaaeb10.json',
    scopes=['https://www.googleapis.com/auth/bigquery']
)

df = pd.read_excel(r"C:\Users\Note reserva\Desktop\Python\arquivos_excel\Produto.xlsx")
df.head()

# Envie o DataFrame para o BigQuery
df.to_gbq(destination_table='pythonengenharia.ProjetoPython.Produto',
          project_id='pythonengenharia',
          if_exists='replace',
          credentials=credencial)






#replace = subistuir a antiga pela original 
#append = adiciona dados no fim da tabela
