==> Banco_de_Dados/06. Google BigQuery/7.Extraindo Varios arquivos.ipynb <==
from google.cloud import bigquery
import pandas as pd

# Especifique o caminho para o arquivo JSON de credenciais
credencial = r"C:\Users\Note reserva\Downloads\pythonengenharia-e4145aaaeb10.json"
# Crie um cliente BigQuery usando as credenciais
client = bigquery.Client.from_service_account_json(credencial)

# querys 
queryProduto = """ 
SELECT * FROM `pythonengenharia.Ecommerce.Produto`
limit 10
"""
queryCategoria=""" 
SELECT * FROM `pythonengenharia.Ecommerce.Categoria`
where id in (0,5,8,10)
"""
queryItens="""
SELECT * FROM `pythonengenharia.Ecommerce.Items`
"""
queryordens="""
SELECT * FROM `pythonengenharia.Ecommerce.Ordens`
limit 20
"""
queryClientes="""
SELECT * FROM `pythonengenharia.Ecommerce.clientes`
limit 50
"""


resultadoProduto    = client.query(queryProduto)
resultadoCategoria  = client.query(queryCategoria)
resultadoItens      = client.query(queryItens)
resultadoOrdens     = client.query(queryordens)
resultadoClientes   = client.query(queryClientes)

# Converta os resultados em um DataFrame
Produto     = resultadoProduto.to_dataframe()
DfCategoria = resultadoCategoria.to_dataframe()
DfItens     = resultadoItens.to_dataframe()
DfOrdens    = resultadoOrdens.to_dataframe()
DfClientes  = resultadoClientes.to_dataframe()

#DfOrdens.head()

# Garantir que a coluna é datetime
DfOrdens['created_at'] = pd.to_datetime(DfOrdens['created_at'])

# Localiza o fuso horário antes de remover
DfOrdens['created_at'] = DfOrdens['created_at'].dt.tz_localize('UTC').dt.tz_convert(None)

pasta = r'C:\Users\Note reserva\Desktop\Curso Engenharia Dados\Python - SGBDS\Extracao\\'
ExcelProduto   = 'Produto.xlsx'
ExcelCategoria ='Categoria.xlsx'
ExcelItens     = 'Itens.xlsx'
ExcelOrdens    = 'Ordens.xlsx'
ExcelClientes  = 'Clientes.xlsx'

Produto.to_excel(pasta+ExcelProduto, index=False)  
DfCategoria.to_excel(pasta+ExcelCategoria, index=False)  
DfItens.to_excel(pasta+ExcelItens, index=False)  
DfOrdens.to_excel(pasta+ExcelOrdens, index=False)  
DfClientes.to_excel(pasta+ExcelClientes, index=False)
